Refactored alcustoms.excel.py into a package with submodules to better maintain the code. Added get_named_cell to alcustoms.excel.Ranges
